+++ lib/ecto/soft_delete_migration.ex
@@ -24,6 +24,6 @@ defmodule Ecto.SoftDelete.Migration do
 
   """
   def soft_delete_columns do
-    add(:deleted_at, :utc_datetime, [])
+    add(:deleted_at, :utc_datetime_usec, [])
   end
 end