+++ test/soft_delete_repo_test.exs
@@ -12,6 +12,14 @@ defmodule Ecto.SoftDelete.Repo.Test do
     end
   end
 
+  defmodule Nondeletable do
+    use Ecto.Schema
+
+    schema "nondeletable" do
+      field(:value, :string)
+    end
+  end
+
   setup do
     :ok = Ecto.Adapters.SQL.Sandbox.checkout(Ecto.SoftDelete.Test.Repo)
   end
@@ -22,7 +30,7 @@ defmodule Ecto.SoftDelete.Repo.Test do
 
       assert %User{} = Repo.soft_delete!(user)
 
-      assert %DateTime{} = Repo.get_by!(User, email: "test0@example.com").deleted_at
+      assert %DateTime{} = Repo.get_by!(User, [email: "test0@example.com"], with_deleted: true).deleted_at
     end
 
     test "should return an error deleting" do
@@ -41,7 +49,7 @@ defmodule Ecto.SoftDelete.Repo.Test do
 
       assert {:ok, %User{}} = Repo.soft_delete(user)
 
-      assert %DateTime{} = Repo.get_by!(User, email: "test0@example.com").deleted_at
+      assert %DateTime{} = Repo.get_by!(User, [email: "test0@example.com"], with_deleted: true).deleted_at
     end
 
     test "should return an error deleting" do
@@ -62,15 +70,42 @@ defmodule Ecto.SoftDelete.Repo.Test do
 
       assert Repo.soft_delete_all(User) == {3, nil}
 
-      assert User |> Repo.all() |> length() == 3
+      assert User |> Repo.all(with_deleted: true) |> length() == 3
 
-      assert %DateTime{} = Repo.get_by!(User, email: "test0@example.com").deleted_at
-      assert %DateTime{} = Repo.get_by!(User, email: "test1@example.com").deleted_at
-      assert %DateTime{} = Repo.get_by!(User, email: "test2@example.com").deleted_at
+      assert %DateTime{} = Repo.get_by!(User, [email: "test0@example.com"], with_deleted: true).deleted_at
+      assert %DateTime{} = Repo.get_by!(User, [email: "test1@example.com"], with_deleted: true).deleted_at
+      assert %DateTime{} = Repo.get_by!(User, [email: "test2@example.com"], with_deleted: true).deleted_at
     end
 
     test "when no results are found" do
       assert Repo.soft_delete_all(User) == {0, nil}
+    end
+  end
+
+  describe "prepare_query/3" do
+    test "excludes soft deleted records by default" do
+      user = Repo.insert!(%User{email: "test0@example.com"})
+      soft_deleted_user = Repo.insert!(%User{email: "deleted@example.com", deleted_at: DateTime.utc_now()})
+      results = User |> Repo.all()
+
+      assert Enum.member?(results, user)
+      refute Enum.member?(results, soft_deleted_user)
+    end
+
+    test "includes soft deleted records if :with_deleted option is present" do
+      user = Repo.insert!(%User{email: "test0@example.com"})
+      soft_deleted_user = Repo.insert!(%User{email: "deleted@example.com", deleted_at: DateTime.utc_now()})
+      results = User |> Repo.all(with_deleted: true)
+
+      assert Enum.member?(results, user)
+      assert Enum.member?(results, soft_deleted_user)
+    end
+
+    test "works with schemas that don't have deleted_at column" do
+      Repo.insert!(%Nondeletable{value: "stuff"})
+      results = Nondeletable |> Repo.all()
+
+      assert length(results) == 1
     end
   end
 end